+++ sdk/transfer/transfer.js
@@ -19,6 +19,8 @@ class Transfer extends Resource {
      * @param bankCode [string]: code of the receiver bank institution in Brazil. If an ISPB (8 digits) is informed, a PIX transfer will be created, else a TED will be issued. ex: '20018183' or '341'
      * @param branchCode [string]: receiver bank account branch. Use '-' in case there is a verifier digit. ex: '1357-9'
      * @param accountNumber [string]: Receiver Bank Account number. Use '-' before the verifier digit. ex: '876543-2'
+     * @param accountType [string, default "checking"]: Receiver bank account type. This parameter only has effect on Pix Transfers. ex: "checking", "savings" or "salary"
+     * @param externalId [string, default null]: url safe string that must be unique among all your transfers. Duplicated external_ids will cause failures. By default, this parameter will block any transfer that repeats amount and receiver information on the same date. ex: "my-internal-id-123456"
      *
      * Parameters (optional):
      * @param tags [list of strings]: list of strings for reference when searching for transfers. ex: ['employees', 'monthly']
@@ -34,8 +36,8 @@ class Transfer extends Resource {
      *
      */
     constructor({
-                    amount, name, taxId, bankCode, branchCode, accountNumber, scheduled, tags, fee, status,
-                    created, updated, transactionIds, id
+                    amount, name, taxId, bankCode, branchCode, accountNumber, accountType, externalId, scheduled,
+                    tags, fee, status, created, updated, transactionIds, id
                 }) {
         super(id);
         this.amount = amount;
@@ -44,6 +46,8 @@ class Transfer extends Resource {
         this.bankCode = bankCode;
         this.branchCode = branchCode;
         this.accountNumber = accountNumber;
+        this.accountType = accountType;
+        this.externalId = externalId;
         this.scheduled = scheduled;
         this.tags = tags;
         this.fee = fee;